"add entry" button only visible on todays date
calendar preloads with todays date and the add entry button is visible.
when user clicks on a date that is not today, the add entry button is not visible (users may not add an entry on previous or future dates).
when user clicks back on today's date, the add entry button becomes visible again.